Extract Vec3 into its own module
Vec3 now has its own functions, the random vector functions. It does not
make sense for it to remain in the ray module. I would have changed the
vec::random module to be an impl on Vec3, but Vec3 is a type synonym and
I don not think that those can have impls.